requirements.txt
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Babel==1.0.0
Flask-SQLAlchemy==2.4.1
Flask-Table @ git+https://github.com/Rohan-vs/flask_table.git@f018cb9b673ba5e4d8c111e00a087688be43c748
Flask-WTF==0.14.3
future==0.18.2
idna==2.8
imageio==2.8.0
imageio-ffmpeg==0.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
moviepy==1.0.3
mutagen==1.44.0
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.0
Pillow==7.1.2
proglog==0.1.9
pygame==1.9.6
pytaglib==1.4.6
pytube3 @ git+https://github.com/Rohan-vs/pytube.git@695f57982dc5ff24ff28b3efe12dbde46d9c568a
pytz==2020.1
requests==2.21.0
selenium==3.14.0
six==1.14.0
SQLAlchemy==1.3.16
tqdm==4.46.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.24.3
Werkzeug==1.0.1
WTForms==2.3.1
